Init moving s3 input to v2 input API
