+++ x-pack/filebeat/input/s3/config.go
@@ -8,24 +8,19 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/elastic/beats/v7/filebeat/harvester"
 	awscommon "github.com/elastic/beats/v7/x-pack/libbeat/common/aws"
 )
 
 type config struct {
-	harvester.ForwarderConfig `config:",inline"`
-	QueueURL                  string              `config:"queue_url" validate:"nonzero,required"`
-	VisibilityTimeout         time.Duration       `config:"visibility_timeout"`
-	AwsConfig                 awscommon.ConfigAWS `config:",inline"`
-	ExpandEventListFromField  string              `config:"expand_event_list_from_field"`
-	APITimeout                time.Duration       `config:"api_timeout"`
+	QueueURL                 string              `config:"queue_url" validate:"nonzero,required"`
+	VisibilityTimeout        time.Duration       `config:"visibility_timeout"`
+	AwsConfig                awscommon.ConfigAWS `config:",inline"`
+	ExpandEventListFromField string              `config:"expand_event_list_from_field"`
+	APITimeout               time.Duration       `config:"api_timeout"`
 }
 
 func defaultConfig() config {
 	return config{
-		ForwarderConfig: harvester.ForwarderConfig{
-			Type: "s3",
-		},
 		VisibilityTimeout: 300 * time.Second,
 		APITimeout:        120 * time.Second,
 	}
