+++ x-pack/filebeat/input/s3/s3_integration_test.go
@@ -7,6 +7,7 @@
 
 package s3
 
+/*
 import (
 	"context"
 	"net/http"
@@ -25,10 +26,12 @@ import (
 	"github.com/aws/aws-sdk-go-v2/service/sqs"
 	"github.com/aws/aws-sdk-go-v2/service/sqs/sqsiface"
 
-	"github.com/elastic/beats/v7/filebeat/channel"
 	"github.com/elastic/beats/v7/filebeat/input"
+	v2 "github.com/elastic/beats/v7/filebeat/input/v2"
 	"github.com/elastic/beats/v7/libbeat/beat"
 	"github.com/elastic/beats/v7/libbeat/common"
+	"github.com/elastic/beats/v7/libbeat/logp"
+	pubtest "github.com/elastic/beats/v7/libbeat/publisher/testing"
 	"github.com/elastic/beats/v7/libbeat/tests/resources"
 	awscommon "github.com/elastic/beats/v7/x-pack/libbeat/common/aws"
 )
@@ -128,7 +131,7 @@ func collectOldMessages(t *testing.T, queueURL string, visibilityTimeout int64, 
 	return oldMessageHandles
 }
 
-func (input *s3Input) deleteAllMessages(t *testing.T, awsConfig awssdk.Config, queueURL string, visibilityTimeout int64, svcSQS sqsiface.ClientAPI) error {
+func (input *s3Collector) deleteAllMessages(t *testing.T, awsConfig awssdk.Config, queueURL string, visibilityTimeout int64, svcSQS sqsiface.ClientAPI) error {
 	var messageReceiptHandles []string
 	init := true
 	for init || len(messageReceiptHandles) > 0 {
@@ -150,33 +153,38 @@ func defaultTestConfig() *common.Config {
 	})
 }
 
-func runTest(t *testing.T, cfg *common.Config, run func(t *testing.T, input *s3Input, out *stubOutleter)) {
-	// Simulate input.Context from Filebeat input runner.
-	inputCtx := newInputContext()
-	defer close(inputCtx.Done)
-
-	// Stub outlet for receiving events generated by the input.
-	eventOutlet := newStubOutlet()
-	defer eventOutlet.Close()
-
-	connector := channel.ConnectorFunc(func(_ *common.Config, _ beat.ClientConfig) (channel.Outleter, error) {
-		return eventOutlet, nil
-	})
-
-	in, err := NewInput(cfg, connector, inputCtx)
+func runTest(t *testing.T, cfg *common.Config, run func(t *testing.T, input *s3Collector, receiver *eventReceiver)) {
+	plugin := Plugin()
+	inp, err := plugin.Manager.Create(cfg)
 	if err != nil {
 		t.Fatal(err)
 	}
-	s3Input := in.(*s3Input)
-	defer s3Input.Stop()
 
-	run(t, s3Input, eventOutlet)
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+
+	client := pubtest.NewChanClient(0)
+	receiver := &eventReceiver{client.Channel}
+	pipeline := pubtest.ConstClient(client)
+	collector, err := inp.(*s3Input).createCollector(v2.Context{
+		Logger:      logp.NewLogger("test"),
+		Cancelation: ctx,
+	}, pipeline)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	run(t, collector, receiver)
 }
 
 func newInputContext() input.Context {
 	return input.Context{
 		Done: make(chan struct{}),
 	}
+}
+
+type eventReceiver struct {
+	ch chan beat.Event
 }
 
 type stubOutleter struct {
@@ -231,7 +239,7 @@ func TestS3Input(t *testing.T) {
 	inputConfig := defaultTestConfig()
 	config := getConfigForTest(t)
 
-	runTest(t, inputConfig, func(t *testing.T, input *s3Input, out *stubOutleter) {
+	runTest(t, inputConfig, func(t *testing.T, input *s3Collector, receiver *eventReceiver) {
 		awsConfig, err := awscommon.GetAWSCredentials(config.AwsConfig)
 		if err != nil {
 
@@ -242,8 +250,8 @@ func TestS3Input(t *testing.T) {
 			s3BucketRegion = "us-west-1"
 		}
 		awsConfig.Region = s3BucketRegion
-		input.awsConfig = awsConfig.Copy()
-		svcSQS := sqs.New(awsConfig)
+		awsConfig = awsConfig.Copy()
+		input.svc := sqs.New(awsConfig)
 
 		// remove old messages from SQS
 		err = input.deleteAllMessages(t, awsConfig, config.QueueURL, int64(config.VisibilityTimeout.Seconds()), svcSQS)
@@ -392,3 +400,4 @@ func TestMockS3Input(t *testing.T) {
 		}
 	})
 }
+*/
